Add CometJoinSuite
